Initial attempt at a koji task based on the Indirection plugin
This creates a new koji task type and CLI option called
image-build-indirection

This task builds an output image using the Image Factory "Indirection"
plugin

It takes as its input two already built images, either as explicit
task numbers of NVR and arch.  It also requires a modified TDL
file as described in the Image Factory source.

The two input images are referred to as the base_image, which contains
content to be turned into another output image, and the utility_image,
which is a full system that will be booted to do the transformation.

The TDL, or "indirection_template" describes the tasks to be performed
inside of the running base image to generate the output.

This feature has the potential to build all of the following types of
images:

Disk Image Builder disk images, built with DIB elements using DIB itself
Live ISO images created using Live Media Creator
Docker layered images built using an arbitrary Docker version and a
Docker base image already built inside of koji